Copy over revisions from the trunk:
------------------------------------------------------------------------
r34036 | etxrab | 2010-09-01 05:39:29 -0700 (Wed, 01 Sep 2010) | 3 lines
Changed paths:
   M /trunk/gtk/voip_calls.c

From Jaap Keuter:
Fix Wireshark don't show mgcp calls in "Telephony" => "VoIP calls"
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5167
------------------------------------------------------------------------
r34037 | guy | 2010-09-01 10:49:27 -0700 (Wed, 01 Sep 2010) | 2 lines
Changed paths:
   M /trunk/gtk/voip_calls.c

Expand and update comments.
------------------------------------------------------------------------
r34038 | etxrab | 2010-09-01 12:42:59 -0700 (Wed, 01 Sep 2010) | 2 lines
Changed paths:
   M /trunk/gtk/voip_calls_dlg.c

1.4 version does not display the t.38 graph.
(init/remove tap_listener missing)
------------------------------------------------------------------------
r34039 | stig | 2010-09-01 13:54:41 -0700 (Wed, 01 Sep 2010) | 2 lines
Changed paths:
   M /trunk/print.c

Ensure we always print the heading when dumping PSML and CSV.
------------------------------------------------------------------------
r34058 | kukosa | 2010-09-03 03:31:56 -0700 (Fri, 03 Sep 2010) | 1 line
Changed paths:
   M /trunk/epan/dissectors/packet-per.c

Fix start offset and length handling in SEQUNCE OF (caused Gtk crash for sequences at the end of frame)
------------------------------------------------------------------------
r34062 | etxrab | 2010-09-03 10:10:52 -0700 (Fri, 03 Sep 2010) | 2 lines
Changed paths:
   M /trunk/epan/dissectors/packet-gtpv2.c

GTPv2: IMSI is decoded improperly.
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5179
------------------------------------------------------------------------
r34069 | stig | 2010-09-06 10:31:17 -0700 (Mon, 06 Sep 2010) | 3 lines
Changed paths:
   M /trunk/epan/dissectors/packet-nas_eps.c

From Pascal Quantin via bug 5186:
Fix the decoding of the maximum/guaranteed bit rates for uplink/downlink.
------------------------------------------------------------------------
r34085 | sfisher | 2010-09-08 13:52:00 -0700 (Wed, 08 Sep 2010) | 6 lines
Changed paths:
   M /trunk/epan/dissectors/packet-tcp.c

Minor correction to SVN revision 34084: we better only try up to the 9
available values in flags[], so we don't run past the end of the array
if one of the reserved flags is set. 
------------------------------------------------------------------------
r34086 | sfisher | 2010-09-08 14:11:45 -0700 (Wed, 08 Sep 2010) | 5 lines
Changed paths:
   M /trunk/epan/dissectors/packet-tcp.c

Make the Windows buildbot happy: guint8 -> guint16 for flags variable in
definition of tcp_analyze_sequence_number(). 
------------------------------------------------------------------------
r34087 | stig | 2010-09-08 14:25:14 -0700 (Wed, 08 Sep 2010) | 2 lines
Changed paths:
   M /trunk/epan/dissectors/packet-tcp.c

Mask out the header length from th_flags.
------------------------------------------------------------------------
r34122 | morriss | 2010-09-15 08:13:40 -0700 (Wed, 15 Sep 2010) | 6 lines
Changed paths:
   M /trunk/wsutil/inet_v6defs.h
   M /trunk/wsutil/libwsutil.def

Fix https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=5160 differently:

On Windows, rename inet_pton() to ws_inet_pton() so that builds on Vista or
later systems (that have inet_pton() natively) will still work on pre-Vista
systems.
------------------------------------------------------------------------
r34129 | morriss | 2010-09-15 14:41:27 -0700 (Wed, 15 Sep 2010) | 1 line
Changed paths:
   M /trunk/wsutil/inet_v6defs.h
   M /trunk/wsutil/libwsutil.def

Also rename inet_ntop() on Windows
------------------------------------------------------------------------
r34101 | rbalint | 2010-09-11 18:13:42 -0700 (Sat, 11 Sep 2010) | 5 lines
Changed paths:
   M /trunk/epan/dissectors/packet-ipx.c
   M /trunk/epan/dissectors/packet-llc.c
   M /trunk/epan/dissectors/packet-ppp.c
   M /trunk/epan/dissectors/packet-sll.c
   M /trunk/epan/dissectors/packet-sll.h

Decode SLL payloads using a dissector table based on sll.ltype.
The idea and the original patch came from Sebastian Reichel
<elektranox@gmail.com> in
http://bugs.debian.org/cgi-bin/bugreport.cgi?bug=594390
------------------------------------------------------------------------

Update the release notes.

svn path=/trunk-1.4/; revision=34413

+++ gtk/tcp_graph.c
@@ -93,7 +93,7 @@ struct segment {
 
 	guint32 th_seq;
 	guint32 th_ack;
-	guint8  th_flags;
+	guint16 th_flags;
 	guint32 th_win;   /* make it 32 bits so we can handle some scaling */
 	guint32 th_seglen;
 	guint16 th_sport;
